Turn the client into a service
This allows additional layers to be applied to the client. Some default layers can then be used (e.g. always add a user-agent header for the outbound request).

The exterior API remains familiar, except that clients must now be mutable, which is consistent with other implementations.

This will allow for future APIs (like a cheaply-cloneable client backed by a channel) which add ergonomic extension points.