feat(body)!: Remove the TryCloneRequest trait from the body module.
This trait was used to help support retries. Unfortunately, given the fragmentation
of the http_body ecosystem, it is not practical nor terribly useful to provide this
trait which we don't even use in hyperdriver itself. This commit removes the trait
and all associated code.

The hyperdriver::Body type still has a `try_clone` method which can be used to support
this use case in the future.