feat!: remove `TransportStream` type.
TransportStream is effectively a crutch to handle TLS and non-TLS streams
in different features. Instead, we can use plain IO types, and just require
that they implement HasTlsConnectionInfo trait. When the TLS feature is
disabled, this trait is blanket implemented for every type.

fix: Fix mock connections and transport to be more consistent.