bugfix: delayed drop could duplicate connections
When delayed drop happens, if the conenction is not in the Connected state, then it would continue to attempt a connection. When the handshake finishes, the Checkout object was not setting the state to Connected, and instead just returning. Fixing this ensures that connections aren’t re-attempted.

Additionally, checkouts would register waiters for themselves, and then wait for their own connection to be ready. To avoid this, checkouts now close their waiter when their connection is ready and they start the registration process. This means they will be skipped in the pass through the waiters in register_connection.