bugfix: TCP address resolver must be clone and replaced in the transport service
The contract of tower::Service requires that the _same_ service that is polled to
readiness is the one that is polled to handle the request. This means that the
address resolver must be cloned and replaced in the transport service, instead of
just cloned - so that the cloned version is left behind, and the ready version
is used to handle the request.