feat: Client pool can delay drop for checkout
When a checkout gets interrupted by a ready / idle connection, the checkout may have done significant work to create the connection. In delayed drop mode, instead of cancelling the connection attempt, the connection will continue in the background and then be added to the pool as a new idle connection.

This can save significant time when future connections begin. The downside is that it requires heap-allocating the future for the connector, but this penalty should be small compared to the typical network latencies observed and the potential savings (100s of ms) by not throwing away completed connection work.