feature: customizable executors
Both clients and servers support customized executors, including local thread executors. There are still some limitations due to the SharedService type converting the wrapped service to dyn Service + Send + Sync. Its not easy to drop the bounds there, but maybe an alternative approach is possible.

As well, bodies are still limited to be Send due to the use of boxed futures. Eliminating the remaining box futures will solve this problem.