feat(client): Flexible APIs to leverage the TCP transport
The client’s TCP connection routine does a bunch of work to set up the TCP socket for HTTP communication. This work is generally not performed by vanilla `tokio::net::TCPStream::connect`, but is performed by hyperdriver’s TCP Transport.

This update makes hyperdriver’s `TCPStream::connect` use the same scheme as the transport under the hood for connection (not for address resolution though).

It also exposes a `.oneshot()` method on transports to connect a transport without using the rest of the hyperdriver infrastructure.