feat: Add generic body type to PoolableConnection
This allows PoolableConnection to imply Connection, meaning that poll_ready() is available whereever PoolableConnection is available. The cost is the proliferation of generics - specifically, the body type generic ends up being explicit on the pool.

This is okay because connections really are generic over body types, and it doesn’t make sense to have a PoolableConnection which is not also a Connection.